Added cart

+++ RashFedProject/shop/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Products(models.Model):
     name = models.CharField(max_length=50, verbose_name='Продукт')
     category = models.ForeignKey("Categories", models.PROTECT, null=True, verbose_name='Категория')
-    price = models.PositiveSmallIntegerField(verbose_name="Цена")
+    price = models.DecimalField(max_digits=10, decimal_places=2, verbose_name="Цена")
     photo = models.ImageField(upload_to="photos/%Y/%m/%d/", verbose_name="Фото", blank=True)
 
     def __str__(self):
